Further refactored the files in the data package
