+++ app/src/main/java/com/example/android/cryptoapp/data/source/local/db/CryptoCurrencyData.kt
@@ -1,4 +1,4 @@
-package com.example.android.cryptoapp.data.local.db
+package com.example.android.cryptoapp.data.source.local.db
 
 import androidx.room.ColumnInfo
 import androidx.room.Entity
